+++ src/pages/nft-item-page/SaleHistoryItem.vue
@@ -71,7 +71,7 @@
 </template>
 
 <script lang="ts" setup>
-import { Payment } from '@/types'
+import { NftPayment, Payment } from '@/types'
 import { Collapse, AppButton } from '@/common'
 import {
   formatFiatAssetFromWei,
@@ -83,9 +83,64 @@ import { CURRENCIES } from '@/enums'
 import { NftDetails } from '@/pages/nft-item-page/NftDetails.vue'
 import { useI18n } from 'vue-i18n'
 
-const props = defineProps<{ historyItem: Payment }>()
+const props = defineProps<{ historyItem: Payment | NftPayment }>()
 
 const { t } = useI18n()
+
+const getSaleBody = () => {
+  if (props.historyItem.erc20_data) {
+    const historyItem = props.historyItem as Payment
+
+    return [
+      {
+        label: t('sale-history-item.buyer-address-lbl'),
+        value: historyItem.payer_address,
+      },
+      {
+        label: t('sale-history-item.token-lbl'),
+        value: historyItem.erc20_data.symbol,
+      },
+      {
+        label: t('sale-history-item.token-amount-lbl'),
+        value: formatAssetFromWei(
+          historyItem.amount,
+          historyItem.erc20_data.decimals,
+        ),
+      },
+      {
+        label: t('sale-history-item.book-link-lbl'),
+        value: historyItem.book_url,
+        isUrl: true,
+      },
+    ] as NftDetails[]
+  } else {
+    const historyItem = props.historyItem as NftPayment
+
+    return [
+      {
+        label: t('sale-history-item.buyer-address-lbl'),
+        value: historyItem.payer_address,
+      },
+      {
+        label: t('sale-history-item.nft-address'),
+        value: historyItem.nft_address,
+      },
+      {
+        label: t('sale-history-item.nft-id'),
+        value: historyItem.nft_id,
+      },
+      {
+        label: t('sale-history-item.floor-price'),
+        value: formatFiatAssetFromWei(historyItem.floor_price, CURRENCIES.USD),
+      },
+      {
+        label: t('sale-history-item.book-link-lbl'),
+        value: historyItem.book_url,
+        isUrl: true,
+      },
+    ] as NftDetails[]
+  }
+}
 
 const saleHeader: NftDetails[] = [
   {
@@ -105,28 +160,7 @@ const saleHeader: NftDetails[] = [
   },
 ]
 
-const saleBody: NftDetails[] = [
-  {
-    label: t('sale-history-item.buyer-address-lbl'),
-    value: props.historyItem.payer_address,
-  },
-  {
-    label: t('sale-history-item.token-lbl'),
-    value: props.historyItem.erc20_data.symbol,
-  },
-  {
-    label: t('sale-history-item.token-amount-lbl'),
-    value: formatAssetFromWei(
-      props.historyItem.amount,
-      props.historyItem.erc20_data.decimals,
-    ),
-  },
-  {
-    label: t('sale-history-item.book-link-lbl'),
-    value: props.historyItem.book_url,
-    isUrl: true,
-  },
-]
+const saleBody: NftDetails[] = getSaleBody()
 </script>
 
 <style lang="scss" scoped>