+++ src/types/api/book.types.ts
@@ -23,6 +23,7 @@ export type Book = JsonApiRecordBase<'books'> & {
   voucher_token: string
   voucher_token_amount: string
   chain_id: number
+  floor_price: string
 }
 
 export type CreateBookResponse = JsonApiRecordBase<'create_signatures'> & {